Power and sample size calculations for a CRT

CRTpower carries out power and sample size calculations for CRTs.

Usage

CRTpower(
  trial = NULL,
  locations = NULL,
  alpha = 0.05,
  desiredPower = 0.8,
  effect = NULL,
  yC = NULL,
  outcome_type = "d",
  sigma2 = NULL,
  phi = 1,
  N = 1,
  ICC = NULL,
  k = NULL,
  sd_h = 0
)

Arguments

trial

dataframe or 'CRTsp' object containing Cartesian coordinates of locations in columns 'x' and 'y'.

locations

numeric: total number of units available for randomization (required if trial is not specified)

alpha

numeric: confidence level

desiredPower

numeric: desired power

effect

numeric: required effect size

yC

numeric: baseline value of outcome

outcome_type

character with options - 'y': continuous; 'n': count; 'e': event rate; 'p': proportion; 'd': dichotomous.

sigma2

numeric: variance of the outcome (required for outcome_type = 'y')

phi

numeric: overdispersion parameter (for outcome_type = 'n' or outcome_type = 'e')

N

numeric: mean of the denominator for proportions (for outcome_type = 'p')

ICC

numeric: Intra-Cluster Correlation

k

integer: number of clusters in each arm (required if trial is not specified)

sd_h

standard deviation of number of units per cluster (required if trial is not specified)

Value

A list of class 'CRTsp' object comprising the input data, cluster and arm assignments, trial description and results of power calculations

Details

Power and sample size calculations are for a two-arm trial using the formulae of Hemming et al, 2011 which use a normal approximation for the inter-cluster variation. For counts or event rate data a quasi–Poisson model is assumed. The functions do not consider any loss in power due to contamination, loss to follow-up etc. If geolocations are not input power and sample size calculations are based on the scalar input parameters. If a trial dataframe or 'CRTsp' object containing a pre-existing randomization is input then the numbers and sizes of clusters are in the input data are used to estimate the power. If buffer zones have been specified then separate calculations are made for the core area and for the full site
The output is an object of class 'CRTsp' containing any input trial data.frame and values for:

The required numbers of clusters to achieve the specified power.

The design effect based on the input ICC.

Calculations of the nominal power (ignoring any bias caused by contamination effects)

Examples

# Example without input geolocations
examplePower = CRTpower(locations = 3000, ICC=0.10, effect=0.4, alpha = 0.05,
    outcome_type = 'd', desiredPower = 0.8, yC=0.35, k = 20, sd_h=5)
#> *** All of geoscale, locations, kappa, mu needed to simulate a new site ***
# Example with input geolocations and randomisation
examplePower = CRTpower(trial = readdata('example_site.csv'), desiredPower = 0.8,
    effect=0.4, yC=0.35, outcome_type = 'd', ICC = 0.05, k = 20)
#> *** Assuming all clusters are the same size ***