CRTspat: Workflow for Cluster Randomised Trials with Spillover

Design, workflow and statistical analysis of Cluster Randomised Trials of (health) interventions where there may be spillover between the arms (see https://thomasasmith.github.io/index.html).

Author

Maintainer: Thomas Smith Thomas-a.Smith@unibas.ch (ORCID) [copyright holder]

Other contributors:

Lea Multerer [contributor]

Mariah Silkey [contributor]