Read example dataset

readdata reads a file from the package library of example datasets

Usage

readdata(filename)

Arguments

filename

name of text file stored within the package

Value

R object corresponding to the text file

Details

The input file name should include the extension (either .csv or .txt). The resulting object is a data frame if the extension is .csv.

Examples

exampleCRT <- readdata('exampleCRT.txt')