Summary description of a "CRTsp" object

summary.CRTsp provides a description of a "CRTsp" object

Usage

# S3 method for class 'CRTsp'
summary(object, maskbuffer = 0.2, ...)

Arguments

object

an object of class "CRTsp" or a data frame containing locations in (x,y) coordinates, cluster assignments (factor cluster), arm assignments (factor arm) and buffer zones (logical buffer), together with any other variables required for subsequent analysis.

maskbuffer

radius of area around a location to include in calculation of areas

...

other arguments used by summary

Value

No return value, write text to the console.

Examples

summary(CRTsp(readdata('exampleCRT.txt')))
#> ===============================CLUSTER RANDOMISED TRIAL ===========================
#> 
#> Summary of coordinates
#> ----------------------
#>                Min.   : 1st Qu.: Median : Mean   : 3rd Qu.: Max.   :
#>       x        -3.20    -1.39    -0.30    -0.07     1.25     5.16   
#>       y        -5.07    -2.84     0.19     0.05     2.49     6.16   
#> nearestDiscord -3.47    -0.49     0.03     0.14     0.73     3.30   
#> 
#> Total area (within  0.2 km of a location) :  27.6 sq.km
#> Total area (convex hull) :  48.2 sq.km
#> 
#> Geolocation of centroid (radians):
#>           latitude:  -0.43 longitude:  34.2 
#> 
#> Locations and Clusters
#> ----------------------                                          -                        
#> Coordinate system                      (x, y)                        
#> Locations:                                                      1181                        
#> Available clusters (across both arms)                           24                        
#>   Per cluster mean number of points                             49.2                        
#>   Per cluster s.d. number of points                             3.9                        
#> Cluster randomization:                      Independently randomized                        
#> No power calculations to report          -                        
#> 
#> Other variables in dataset
#> --------------------------          RDT_test_result  base_denom  propensity  denom  num